+++ dynamite/config_reader.py
@@ -774,7 +774,6 @@ class Configuration(object):
         self.logger.info('Config file copied to '
                          f'{dest_directory}{self.config_file_name}.')
 
-
     def backup_config_file(self, reset=False, keep=None, delete_other=False):
         """
         Copy the config file to the output directory.
@@ -847,6 +846,10 @@ class Configuration(object):
         """
         Validates the system and settings.
 
+        This includes aligning the number of gh coefficients with the config
+        file setting `number_GH` for Gauss Hermite kinematics and also applies
+        the weight solver settings' systematic errors to the kinematics.
+
         This method is still VERY rudimentary and will be adjusted as we add new
         functionality to dynamite. Currently, this method is geared towards
         legacy mode.
@@ -856,6 +859,8 @@ class Configuration(object):
         None.
 
         """
+        self.settings.validate()
+        _ = self.validate_chi2()
         if sum(1 for i in self.system.cmp_list \
                if isinstance(i, physys.Plummer)) != 1:
             self.logger.error('System must have exactly one Plummer object')
@@ -896,22 +901,17 @@ class Configuration(object):
                         check_gh = (kin_data.type == 'GaussHermite')
                         check_bl = (kin_data.type == 'BayesLOSVD')
                         if (not check_gh) and (not check_bl):
-                            self.logger.error('VisibleComponent kinematics type'
-                                              'must be GaussHermite or '
-                                              'BayesLOSVD')
-                            raise ValueError('VisibleComponent kinematics type'
-                                             'must be GaussHermite or '
-                                             'BayesLOSVD')
+                            txt = 'VisibleComponent kinematics type must be ' \
+                                  'GaussHermite or BayesLOSVD'
+                            self.logger.error(txt)
+                            raise ValueError(txt)
                         if check_bl:
                             # check weight solver type
                             if ws_type == 'LegacyWeightSolver':
-                                self.logger.error("LegacyWeightSolver can't be "
-                                                  "used with BayesLOSVD - use "
-                                                  "weight-solver type NNLS")
-                                raise ValueError("LegacyWeightSolver can't be "
-                                                  "used with BayesLOSVD - use "
-                                                  "weight-solver type NNLS")
-
+                                txt = "LegacyWeightSolver can't be used with "\
+                                      "BayesLOSVD - use weight-solver type NNLS"
+                                self.logger.error(txt)
+                                raise ValueError(txt)
                 else:
                     self.logger.error('VisibleComponent must have kinematics: '
                                       'either GaussHermite or BayesLOSVD')
@@ -990,8 +990,6 @@ class Configuration(object):
                     max_bins += 1
                 for k in stars.kinematic_data:
                     k.hist_bins = max_bins
-        self.settings.validate()
-        _ = self.validate_chi2()
 
     def validate_chi2(self, which_chi2=None):
         """